Serving web server pages from redis cache.
There is a script to load a local redis db with only the temporal data
streams fetched from Socrata. Run that first, and then the web.py server
serves its list of streams from the redis server instead of going live
to the Socrata catalog.

This allows the web server to online display temporal data streams, so
the frontend can assume that every stream has is temporal, making it
easier for me to continue the next phase of client-side development.